Múltiples mejoras

+++ app/src/main/java/com/example/timebank/OfferCreate.java
@@ -9,15 +9,15 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.CheckBox;
 import android.widget.EditText;
+import android.widget.Toast;
 
 import com.parse.ParseObject;
 import com.parse.ParseUser;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.time.format.DateTimeFormatter;
-import java.util.Calendar;
+import java.util.Arrays;
 import java.util.Date;
 
 public class OfferCreate extends AppCompatActivity {
@@ -51,6 +51,8 @@ public class OfferCreate extends AppCompatActivity {
 
     public void saveChanges(View v) {
 
+        String[] categoriesList = getIntent().getStringArrayExtra("categories");
+
         ParseObject offer = new ParseObject("Offer");
         try {
             long expiryDate = getIntent().getExtras().getLong("task_date");
@@ -69,25 +71,50 @@ public class OfferCreate extends AppCompatActivity {
                 .toInstant().toEpochMilli();
         Date date2 = new Date(millis);
 
-        offer.put("expires_at", date2);
-        offer.put("username", ParseUser.getCurrentUser());
-        offer.put("offer_title", offerTitle.getText().toString());
-        offer.put("offer_desc", offerDesc.getText().toString());
-        offer.put("one_time_only", oneTime.isChecked());
-        offer.put("is_finished", false);
 
-        offer.saveInBackground();
-        Intent intent = new Intent(OfferCreate.this, ContentMain.class);
-        sharedpreferences = getSharedPreferences("createOfferTemp", Context.MODE_PRIVATE);
-        SharedPreferences.Editor editor = sharedpreferences.edit();
-        editor.clear();
-        editor.apply();
+        if (offerTitle.getText().toString().equals("") || offerDesc.getText().toString().equals("")) {
+            Toast.makeText(this, "El título o la descripción de la oferta están vacíos", Toast.LENGTH_SHORT).show();
+        } else {
+            offer.put("expires_at", date2);
+            offer.put("username", ParseUser.getCurrentUser());
+            offer.put("offer_title", offerTitle.getText().toString());
+            offer.put("offer_desc", offerDesc.getText().toString());
+            offer.put("one_time_only", oneTime.isChecked());
+            offer.put("community", ParseUser.getCurrentUser().getString("community_name"));
+            offer.put("times_completed", 0);
+            offer.put("is_finished", false);
+            if (categoriesList != null) {
+                offer.put("categories", Arrays.asList(categoriesList));
+            } else {
+                offer.put("categories", Arrays.asList("ninguna"));
+            }
 
-        startActivity(intent);
+            offer.saveInBackground();
+            Intent intent = new Intent(OfferCreate.this, MainMenu.class);
+            sharedpreferences = getSharedPreferences("createOfferTemp", Context.MODE_PRIVATE);
+            SharedPreferences.Editor editor = sharedpreferences.edit();
+            editor.clear();
+            editor.apply();
+
+            startActivity(intent);
+        }
     }
 
     public void cancel(View v) {
-        Intent intent = new Intent(OfferCreate.this, ContentMain.class);
+        Intent intent = new Intent(OfferCreate.this, MainMenu.class);
+        startActivity(intent);
+    }
+
+    public void configCat(View v) {
+        Intent intent = new Intent(OfferCreate.this, Categories.class);
+        intent.putExtra("taskType", "offer");
+        sharedpreferences = getSharedPreferences("createOfferTemp", Context.MODE_PRIVATE);
+        SharedPreferences.Editor editor = sharedpreferences.edit();
+        editor.putString("offerTitle", offerTitle.getText().toString());
+        editor.putString("offerDesc", offerDesc.getText().toString());
+        editor.putBoolean("offerOneTime", oneTime.isChecked());
+        editor.apply();
+        intent.putExtra("TaskCreated", "offer");
         startActivity(intent);
     }
 
@@ -99,6 +126,7 @@ public class OfferCreate extends AppCompatActivity {
         editor.putString("offerDesc", offerDesc.getText().toString());
         editor.putBoolean("offerOneTime", oneTime.isChecked());
         editor.apply();
+        intent.putExtra("TaskCreated", "offer");
         startActivity(intent);
     }
 }
